fix: messy issues with nodeVersionUtils helper needing semver installed
