+++ __tests__/lib/isSupportedNodeVersion.test.js
@@ -1,6 +1,4 @@
-const { isSupportedNodeVersion, getNodeVersion } = require('../../src/lib/nodeVersionUtils');
-const pkg = require('../../package.json');
-const semver = require('semver');
+const isSupportedNodeVersion = require('../../src/lib/isSupportedNodeVersion');
 
 describe('#isSupportedNodeVersion()', () => {
   it('should return true for a supported version of node', () => {
@@ -12,13 +10,5 @@ describe('#isSupportedNodeVersion()', () => {
     expect(isSupportedNodeVersion('10.0.0')).toBe(false);
     expect(isSupportedNodeVersion('12.0.0')).toBe(false);
     expect(isSupportedNodeVersion('18.0.0')).toBe(false);
-  });
-});
-
-describe('#getNodeVersion()', () => {
-  it('should extract version that matches range in package.json', () => {
-    const version = parseInt(getNodeVersion(), 10);
-    const cleanedVersion = semver.valid(semver.coerce(version));
-    expect(semver.satisfies(cleanedVersion, pkg.engines.node)).toBe(true);
   });
 });
